Move schema to separate storage key
Big schemas (megabytes in SDL) take a few hundred milliseconds to serialize, and doing this on every keystroke is wasteful and causes sluggish performance. By moving it to its own key, the equals-check that is already in place ensures that the schema is only written when it changed.

This change is backwards-compatible as it still reads the schema from the main window key while there is no separate schema key present.