Added altair-graphql-core.
Removed altair-exported-types.
altair-exported-types was a temporary hack to getting Altair types
available for the other packages like altair-static. As it turns out
exporting types only from a typescript project introduces other issues
since the type implementations are not exported as well. Things like
enums do not get typed properly when used and they need the provided
strings to be recast as the enum type which defeats the purpose of
having the types in the first place. For example, `PluginSource` which
includes `NPM` would now have to be typed as `'NPM' as PluginSource`
which also means you do not get the autocomplete options to ensure the
string you are passing is indeed a valid value for that enum.
This also affects classes as well which would no longer be extendable
but can only be implemented as an interface.