Fix proxy settings issue (#1948)
Currently, there is a bug where it is impossible to set the proxy settings via the `Desktop Settings` menu. This is because `initialData` start as `undefined`, thus causing `Object.keys()` to throw an error and the rest of the initializer function never runs. This means that none of the save & submit functionality works because it never gets setup properly.

To fix this, you can just default `initialData` to an empty object.